WIP - We don't actually need the "ID" version of the code error component
